progress: 46.70% of code matched
Source-commit: 9e7fca6aa76802d1e9ac9798a2fece5b18274f94

+++ src/game/game_fn_801383B4.c
@@ -59,7 +59,7 @@ void fn_801383B4(int slot_index, int mode)
     header = (ResourceHeader*)scratch;
     if (read_size == 0x20) {
         parent = header->parent;
-    } else if (read_size != 0) {
+    } else if ((int)read_size != 0) {
         count = header->count;
         data_offset = 0;
         data_index = 0;
